+++ apps/catalog/catalog/src/app/marketplace/title/list/list.component.scss
@@ -32,4 +32,9 @@ list-filter {
   justify-content: center;
 }
 
+pdf-export {
+  button {
+    margin: 24px;
+  }
+}
